+++ src/main/java/dev/jaxydog/astral/content/effect/CustomPotions.java
@@ -14,9 +14,7 @@ public final class CustomPotions extends ContentRegistrar {
 
     public static final CustomPotion SINISTER = new CustomPotion(
         "sinister",
-        new Recipe(Potions.THICK,
-            Ingredient.ofItems(AstralItems.DYEABLE_AMETHYST_CLUSTERS.get(DyeColor.RED).orElseThrow())
-        ),
+        new Recipe(Potions.THICK, Ingredient.ofItems(AstralItems.DYEABLE_AMETHYST_CLUSTERS.getComputed(DyeColor.RED))),
         new StatusEffectInstance(CustomStatusEffects.SINISTER, 15 * 20, 0)
     );