+++ src/main/java/dev/jaxydog/astral/register/RegistrationPriority.java
@@ -1,0 +1,46 @@
+/*
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ *
+ * Copyright © 2024 Jaxydog
+ *
+ * This file is part of Astral.
+ *
+ * Astral is free software: you can redistribute it and/or modify it under the terms of the GNU Affero General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
+ *
+ * Astral is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License along with Astral. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package dev.jaxydog.astral.register;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * Instructs the registration system on the ordering with which to register this value as compared to other fields.
+ * <p>
+ * By default, fields without this annotation are assumed to have a priority of {@code 0}.
+ *
+ * @author Jaxydog
+ * @see ContentRegistrar
+ * @since 2.0.0
+ */
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.FIELD)
+public @interface RegistrationPriority {
+
+    /**
+     * The priority of this field.
+     * <p>
+     * Higher priority means that the field will be registered sooner.
+     *
+     * @return The priority.
+     *
+     * @since 2.0.0
+     */
+    int value();
+
+}